Struct logs_trait::MyLogsProcessor

pub(crate) struct MyLogsProcessor {
    pub(crate) counter: Arc<AtomicUsize>,
}

Custom log processor that increments a counter for each log record.

This is a simple example of a custom log processor that can be used to count the number of log records that are processed.

This needs to derive Clone (and store the counter in an Arc) as the runtime could need multiple Services to process the logs.

Fields
counter: Arc<AtomicUsize>
Implementations

impl MyLogsProcessor

pub fn new() -> Self